// frontend/lib/api/generated/model/centroConsumoOut.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SIGA-UNMSM API
 * API del Sistema Integral de Gestión de Almacén y Abastecimiento Alimentario — UNMSM. Módulos activos: Autenticación, Usuarios, Almacenes, Catálogo Nutricional.
 * OpenAPI spec version: 0.1.0
 */

export interface CentroConsumoOut {
  centro_consumo_id: number;
  sede_id: number;
  almacen_id: number;
  nombre: string;
  poblacion_referencia: number | null;
}